+++ eq-author/src/components/ContentPickerv2/DestinationPicker/index.js
@@ -45,7 +45,7 @@ const DestinationPicker = ({ data, onSelected, isSelected, ...otherProps }) => (
 DestinationPicker.propTypes = {
   data: PropTypes.shape({
     pages: PropTypes.array,
-    logicalDestinations: PropTypes.array,
+    logicalDestinations: PropTypes.func,
     sections: PropTypes.array,
   }),
   onSelected: PropTypes.func.isRequired,
