+++ eq-author/src/constants/destinationKey.js
@@ -1,0 +1,11 @@
+export const EndOfQuestionnaire = "EndOfQuestionnaire";
+export const NextPage = "NextPage";
+export const Default = "Default";
+export const EndOfCurrentSection = "EndOfCurrentSection";
+
+export const destinationKey = {
+  [EndOfQuestionnaire]: "End of questionnaire",
+  [EndOfCurrentSection]: "End of current section",
+  [NextPage]: "Next page",
+  [Default]: "Select a destination",
+};
